#681 #611, add changes related to pull request comments
