+++ TypeCobol/Compiler/Parser/ProgramClassParserStep.cs
@@ -87,7 +87,8 @@ namespace TypeCobol.Compiler.Parser
             //Create link between datas
             programClassBuilder.SyntaxTree.Root.AcceptASTVisitor(new TypeCobolLinker());
             //Complete some information on Node and run checker that need a full AST
-            programClassBuilder.SyntaxTree.Root.AcceptASTVisitor(new Cobol85CompleteASTChecker(diagnostics));
+            programClassBuilder.SyntaxTree.Root.AcceptASTVisitor(new Cobol85CompleteASTChecker());
+            programClassBuilder.SyntaxTree.Root.AcceptASTVisitor(new DiagnosticsChecker(diagnostics));
 
 
             //Stop measuring tree building performance
